Commented out unused code in RosteringImporter.java; Updated OptaPlanner to 7.35, Apache POI to 4.1.2, PDFBox to 2.0.19

+++ Quoll/src/main/java/au/org/thebigissue/rostering/solver/solution/Roster.java
@@ -208,7 +208,7 @@ public class Roster extends AbstractPersistable {
 
         variance=variance/(hoursByStaff.size());
 
-        return (int) Math.round(variance);
+        return Math.round(variance);
     }
 
     public int calculateFacilitatorVariance() {
